Add 'evidence' fixture
This fixture allows attachment of an evidence (a file) to a Test Run,
with one line of code in the test function.

It can take a filename, data and content-type string as arguments. The
result depends on the presence and value of each of these arguments.
See the table in the comments of function 'wrapper_evidence' in file
src/pytest_xray/plugin.py.

Tests are included to cover 100% of the added code.